+New Dice Button, much improved button animations
I redid the animations somewhat, losing the empty parent objects that
weren't actually helping this time, so things are a little cleaner now.
Also removed the old Doors script I brought in as reference for
animating objects. Also also changed the d10 prefab in the library (it
had two "d 10" scripts on it - now it has only one).