Pretty Die Edit Menu, Lookpoints for Camera
The new Die Edit menu has buttons and slider, which do nothing, except
the NOPE button which tells the camera to FlyTo() the lookpoint for
Dicebox. On spawn, dice are told who the camera is, so tapholding them
will tell camera to FlyTo() the Die Edit lookpoint.
Done at 2:00-3:00 AM, so I apologize for whatever this breaks. @u@